+++ src/FB/Monitor.h
@@ -55,7 +55,7 @@ protected:
   Vec3r g_box;           ///< グローバル領域サイズ
   int* bid;              ///< 境界ID
   int* bcd;              ///< BCindex B
-  float* cut;            ///< 交点情報
+  long long* cut;        ///< 交点情報
   TextParser* tpCntl;    ///< テキストパーサへのポインタ
   REAL_TYPE area;        ///< 断面積 [m^2]
   int NoCompo;           ///< 物性テーブルの個数 >> 配列の大きさは[NoCompo+1]
@@ -207,7 +207,7 @@ public:
   ///   @param [in] m_mtbl         物性テーブル
   ///
   void setControlVars(int* bid,
-                      float* cut,
+                      long long* cut,
                       int* bcd,
                       const REAL_TYPE refVelocity,
                       const REAL_TYPE baseTemp,
